# src/main/resources/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/accounts/{id}":{"get":{"tags":["account-resource"],"summary":"Get account by ID","description":"Retrieve an account by its ID","operationId":"findById","parameters":[{"name":"id","in":"path","description":"ID of the account to be retrieved","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successfully retrieved account","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}},"404":{"description":"Account not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}}},"put":{"tags":["account-resource"],"summary":"Update an existing account","description":"Update the details of an existing account","operationId":"update","parameters":[{"name":"id","in":"path","description":"ID of the account to be updated","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"currency","in":"query","description":"New currency code of the account","required":true,"schema":{"type":"string","maxLength":3,"minLength":3}},{"name":"balance","in":"query","description":"New balance of the account","required":true,"schema":{"type":"number","minimum":0.0}}],"responses":{"200":{"description":"Successfully updated account","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}},"409":{"description":"Account already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}},"502":{"description":"Currency service unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}}},"delete":{"tags":["account-resource"],"summary":"Delete an account","description":"Delete an account by its ID","operationId":"delete","parameters":[{"name":"id","in":"path","description":"ID of the account to be deleted","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successfully deleted account"}}}},"/transactions":{"post":{"tags":["transaction-resource"],"summary":"Create a new transaction","description":"Create a new transaction with the given details","operationId":"create","parameters":[{"name":"sourceAccountId","in":"query","description":"ID of the account initiating the transfer","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"targetAccountId","in":"query","description":"ID of the account receiving the transfer","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"amount","in":"query","description":"The amount transferred in the initiator currency","required":true,"schema":{"type":"number","minimum":0.0}}],"responses":{"400":{"description":"Insufficient balance"},"200":{"description":"Successfully created transaction"},"502":{"description":"Currency service unavailable"}}}},"/accounts":{"get":{"tags":["account-resource"],"summary":"Get all accounts","description":"Retrieve a list of all accounts","operationId":"findAll","responses":{"200":{"description":"Successfully retrieved list","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}}}},"post":{"tags":["account-resource"],"summary":"Create a new account","description":"Create a new account with the given details","operationId":"create_1","parameters":[{"name":"id","in":"query","description":"ID of the new account","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"currency","in":"query","description":"Currency code of the new account","required":true,"schema":{"type":"string","maxLength":3,"minLength":3}},{"name":"balance","in":"query","description":"Initial balance of the new account","required":true,"schema":{"type":"number","minimum":0.0}}],"responses":{"200":{"description":"Successfully created account","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}},"409":{"description":"Account already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}},"502":{"description":"Currency service unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}}}}},"components":{"schemas":{"Account":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"currency":{"type":"string"},"balance":{"type":"number"}}}}}}